+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,9 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <flutter_ime/flutter_ime_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  FlutterImePluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FlutterImePluginCApi"));
 }